backend/requirements.txt
google-generativeai>=0.3.0
aiohttp>=3.9.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
sqlalchemy>=2.0.23
alembic>=1.12.1
pytest>=7.4.3
httpx>=0.25.2
python-dotenv>=1.0.0
rich>=13.7.0
stripe>=7.0.0

